functions moved to utils
